third_party/blink/web_tests/external/wpt/html/browsers/the-window-object/window-open-noopener_indexed-expected.txt

This is a testharness.js-based test.
PASS window.open() with 'noopener' should not reuse existing target
PASS noopener=1 means the same as noopener
FAIL noopener=0 means lack of noopener assert_equals: expected object "[object Window]" but got null
FAIL noopener needs to be present as a token on its own assert_equals: expected object "[object Window]" but got null
PASS Trailing noopener should work
PASS Leading noopener should work
PASS Interior noopener should work
Harness: the test ran to completion.
